+++ CHANGES.md
@@ -2,7 +2,7 @@
 
 ## [5.3a0] - 21 May 2025
 - add API to get name from ORCID
-- add identifiers to Person
+- add ORCID Person
 
 ## [5.2] - 16 May 2025
 - add search by author_ids